Fix bug not setting the opaque for SSL socket. Attach now returns the status code, not new socket.
git-svn-id: https://svn.apache.org/repos/asf/tomcat/connectors/trunk/jni@300856 13f79535-47bb-0310-9956-ffa450edef68

+++ java/org/apache/tomcat/jni/SSLSocket.java
@@ -33,9 +33,9 @@ public class SSLSocket {
      * @param sock APR Socket that already did physical connect.
      * @param pool The pool to use
      * @param pool The pool to use
-     * @return The new socket that has been set up.
+     * @return APR_STATUS code.
      */
-    public static native long attach(long ctx, long sock, long pool)
+    public static native int attach(long ctx, long sock, long pool)
         throws Exception;
 
     /**
